Added clarification for release badge
Since this readme will also be on PyPi, it can confuse people when selecting a specific release and seeing a newer version for release. 
Although when uploading to PyPi, the readme is the one that was active at that moment, it still fetches the image from shields.io which will update with new releases.